Bug fixes and BUILD!
Fixed build errors
Successfully built application

Must fix trusted domain issue though
