+++ src/server/auth/config.ts
@@ -41,6 +41,7 @@ export const authConfig = {
      *
      */
   ],
+  // @ts-expect-error - Fix later
   adapter: PrismaAdapter(db),
   callbacks: {
     session: ({ session, user }) => ({
